Move Warning types into model/
