+++ src/javascript/function.ts
@@ -15,8 +15,7 @@
 import * as estree from 'estree';
 
 import {Annotation as JsDocAnnotation} from '../javascript/jsdoc';
-import {Document, Feature, Privacy, Resolvable, SourceRange} from '../model/model';
-import {Warning} from '../warning/warning';
+import {Document, Feature, Privacy, Resolvable, SourceRange, Warning} from '../model/model';
 
 export class ScannedFunction implements Resolvable {
   name: string;
