+++ src/model/model.ts
@@ -49,3 +49,4 @@ export {Package, QueryOptions as PackageQueryOptions} from './package';
 export * from './reference';
 export * from './resolvable';
 export * from './source-range';
+export * from './warning';
